pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vbai"
version = "0.1.0"
description = "Visual Brain AI - Multi-task brain MRI analysis library"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "HealFuture", email = "contact@healfuture.com"}
]
maintainers = [
    {name = "HealFuture", email = "contact@healfuture.com"}
]
keywords = [
    "deep-learning",
    "medical-imaging",
    "brain-mri",
    "dementia",
    "tumor-detection",
    "pytorch",
    "multi-task-learning"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Image Processing",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "Pillow>=8.0.0",
    "numpy>=1.19.0",
]

[project.optional-dependencies]
full = [
    "matplotlib>=3.3.0",
    "opencv-python>=4.5.0",
    "tqdm>=4.60.0",
    "tensorboard>=2.4.0",
    "pyyaml>=5.4.0",
]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.0.0",
    "black>=21.0",
    "isort>=5.0.0",
    "flake8>=3.9.0",
    "mypy>=0.900",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=0.5.0",
    "sphinx-autodoc-typehints>=1.12.0",
]
all = [
    "vbai[full,dev,docs]",
]

[project.urls]
Homepage = "https://github.com/healfuture/vbai"
Documentation = "https://vbai.readthedocs.io"
Repository = "https://github.com/healfuture/vbai"
Issues = "https://github.com/healfuture/vbai/issues"

[project.scripts]
vbai-train = "vbai.cli:train_cli"
vbai-predict = "vbai.cli:predict_cli"

[tool.setuptools]
packages = ["vbai", "vbai.models", "vbai.data", "vbai.training", "vbai.utils", "vbai.configs"]

[tool.setuptools.package-data]
vbai = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=vbai --cov-report=term-missing"
